requirements.txt
streamlit
beautifulsoup4
requests
textblob
pandas
Pillow
plotly
torch
ftfy
transformers
scikit-learn
numpy